move new tests to new file and minor updates based on feedback
move new tests to new file and minor updates based on feedback

Add tags to README (#20)

bump patch number
