+++ lib/resValidator.js
@@ -73,7 +73,6 @@ function parseOpenApiFile(filePath) {
 
 function findResExpectedByApiSpec(res, openApiSpec) {
   const openApiPath = getOAPath(res.req, openApiSpec);
-
   const pathObj = openApiSpec.paths[openApiPath];
   if (!pathObj) {
     throw new Error(`No '${openApiPath}' path defined in OpenAPI spec`);
@@ -111,7 +110,7 @@ function validateResAgainstApiSpec(res, expectedResponse, openApiSpec) {
 
 
 function getOAPath(req, openApiSpec) {
-  let pathname = url.parse(req.path).pathname; // excludes the query (because: path = pathname + query)
+  const pathname = url.parse(req.path).pathname; // excludes the query (because: path = pathname + query)
 
   const serverUrl = getServerUrlUsed(pathname, openApiSpec.servers);
 
@@ -119,17 +118,16 @@ function getOAPath(req, openApiSpec) {
   if (openApiSpec.servers && !serverUrl) {
     throw new Error(`No server matching '${pathname}' path defined in OpenAPI spec`);
   }
-  pathname = removeServerURLFromPath(pathname, serverUrl);
-
-  if (openApiSpec.paths.hasOwnProperty(pathname)) {
-    return pathname;
+  const pathnameWithoutServerUrl = removeServerURLFromPath(pathname, serverUrl);
+  if (openApiSpec.paths.hasOwnProperty(pathnameWithoutServerUrl)) {
+    return pathnameWithoutServerUrl;
   }
-  const OAPaths = Object.keys(openApiSpec.paths);
-  const OAPath = findOAPathMatchingPathname(pathname, OAPaths);
-  return OAPath || pathname;
+  const OAPath = findOAPathMatchingPathname(pathnameWithoutServerUrl, openApiSpec);
+  return OAPath || pathnameWithoutServerUrl;
 }
 
-function findOAPathMatchingPathname(pathname, OAPaths) {
+function findOAPathMatchingPathname(pathname, openApiSpec) {
+  const OAPaths = Object.keys(openApiSpec.paths);
   const OAPathsWithPathParams = OAPaths.filter(path => path.includes('{'));
   const OAPath = OAPathsWithPathParams.find(OAPath => {
     const pathInColonForm = OAPath.replace(/{/g, ':').replace(/}/g, ''); // converts all {foo} to :foo
@@ -140,14 +138,14 @@ function findOAPathMatchingPathname(pathname, OAPaths) {
   return OAPath;
 }
 
-function getServerUrlUsed(openApiFullPath, servers) {
+function getServerUrlUsed(openApiPathWithServerUrl, servers) {
   if (!servers) {
     return null;
   }
-  const serverUsed = servers.find(server => openApiFullPath.startsWith(server.url));
+  const serverUsed = servers.find(server => openApiPathWithServerUrl.startsWith(server.url));
   return serverUsed ? serverUsed.url : null;
 }
 
-function removeServerURLFromPath(openApiFullPath, serverUrlUsed) {
-  return openApiFullPath.replace(serverUrlUsed, '');
+function removeServerURLFromPath(openApiPathWithServerUrl, serverUrlUsed) {
+  return openApiPathWithServerUrl.replace(serverUrlUsed, '');
 }
